feat: Refactor Text Component
