+++ src/components/ConfirmationScreen/CancelConfirmation/CancelConfirmation.tsx
@@ -1,6 +1,7 @@
 import styled from '@emotion/styled/macro';
 import { Button, ButtonKind } from '../Button/Button';
-import { Text, TextKind } from '../Text/Text';
+import { Text } from '../Text/Text';
+import { TextKind } from '../Text/TextTheme';
 
 export interface CancelConfirmationProps {
   onBack: () => void;
