+++ src/components/ConfirmationScreen/ConfirmSwap/ConfirmSwap.tsx
@@ -2,7 +2,8 @@ import styled from '@emotion/styled/macro';
 import { AssetInput, AssetInputProps } from '../AssetInput/AssetInput';
 import { Button, ButtonKind, ButtonPadding } from '../Button/Button';
 import { Table, TableProps } from '../Table/Table';
-import { Text, TextKind } from '../Text/Text';
+import { Text } from '../Text/Text';
+import { TextKind } from '../Text/TextTheme';
 import { RefreshTimer } from '../RefreshTimer/RefreshTImer';
 import { Icon } from '../Icon/Icon';
 import { ErrorMessage } from '../ErrorMessage/ErrorMessage';
