+++ src/components/ConfirmationScreen/RefreshTimer/RefreshTImer.tsx
@@ -1,6 +1,7 @@
 import { keyframes } from '@emotion/react';
 import styled from '@emotion/styled/macro';
-import { Text, TextKind } from '../Text/Text';
+import { Text } from '../Text/Text';
+import { TextKind } from '../Text/TextTheme';
 
 export interface RefreshTimerProps {
   time: number;
